Splitted, feature added
Missing comments and work days
Missing get day.
